data_capturing_and_measurement/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
dnspython==2.4.2
